refactor: move Cars and Round test classes to race package
