+++ src/test/kotlin/racingcar/domain/race/RoundTest.kt
@@ -1,9 +1,8 @@
-package racingcar.domain
+package racingcar.domain.race
 
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.domain.race.Round
 
 class RoundTest {
